Clarify debmake issue due to lack of shebang in setup.py
This patch updates the README file of this project to give a valid
workaround and explain the root cause of the problem. A valid patch will
be sent to the debmake community to address cases where the setup.py does
not have a shebang. In collaboration with the debamake community, we will
try to arrive to the best solution.

Signed-off-by: VictorRodriguez <vm.rod25@gmail.com>